Adapt delta quadplane to standard wing model
