+++ Tools/parametric_model/src/models/aerodynamic_models/__init__.py
@@ -5,7 +5,6 @@ __maintainer__ = "Manuel Galliker"
 __license__ = "BSD 3"
 
 
-from .aero_model_Delta import AeroModelDelta
 from .fuselage_drag_model import FuselageDragModel
 from .elevator_model import ElevatorModel
 from .standard_wing_model import StandardWingModel
