+++ Tools/parametric_model/src/models/aerodynamic_models/standard_wing_model.py
@@ -76,7 +76,47 @@ class StandardWingModel():
 
         return X_wing_body_frame
 
-    def compute_aero_features(self, v_airspeed_mat, angle_of_attack_vec):
+    def compute_control_surface_force_features(self, v_airspeed, angle_of_attack, flap_commands):
+        """
+        Model description:
+        The model assumes that the ailerons will always be 
+
+
+        flap_commands: TBD
+        1: aileron command left/right?
+        2. aileron command left/right?
+        """
+
+        v_xz = math.sqrt(v_airspeed[0]**2 + v_airspeed[2]**2)
+        X_cs = np.zeros((3, 2))
+
+        # aileron drag
+        X_cs[0, 0] = -(abs(flap_commands[0]) +
+                       abs(flap_commands[1]))*v_xz**2
+        X_cs[0, 1] = -(abs(flap_commands[0]) +
+                       abs(flap_commands[1]))**2*v_xz**2
+
+        R_aero_to_body = Rotation.from_rotvec(
+            [0, -angle_of_attack, 0]).as_matrix()
+        X_rudder_body = R_aero_to_body @ X_cs
+        return X_rudder_body
+
+    def compute_single_aero_feature(self, v_airspeed, angle_of_attack, flap_commands):
+        """
+        Used to compute the feature matrices for a single timestamp
+
+        Inputs:
+
+        v_airspeed: numpy array of dimension (1,3) with [v_a_x, v_a_y, v_a_z]
+        angle_of_attack: corresponding AoA values
+        roll_commands = numpy array of dimension (1,3) with columns for [u_roll_1, u_roll_2, u_pitch_collective]
+        """
+        X_wing = self.compute_wing_force_features(v_airspeed, angle_of_attack)
+        X_cs = self.compute_control_surface_force_features(
+            v_airspeed, angle_of_attack, flap_commands)
+        return np.hstack((X_wing, X_cs))
+
+    def compute_aero_features(self, v_airspeed_mat, angle_of_attack_vec, flap_commands):
         """
         Inputs:
 
@@ -85,17 +125,19 @@ class StandardWingModel():
         flap_commands = numpy array of dimension (n,3) with columns for [u_aileron_right, u_aileron_left, u_elevator]
         """
         print("Starting computation of aero features...")
-        X_aero = self.compute_wing_force_features(
-            v_airspeed_mat[0, :], angle_of_attack_vec[0])
+        X_aero = self.compute_single_aero_feature(
+            v_airspeed_mat[0, :], angle_of_attack_vec[0], flap_commands[0, :])
         aero_features_bar = Bar(
             'Feature Computatiuon', max=v_airspeed_mat.shape[0])
         for i in range(1, len(angle_of_attack_vec)):
-            X_curr = self.compute_wing_force_features(
-                v_airspeed_mat[i, :], angle_of_attack_vec[i])
+            X_curr = self.compute_single_aero_feature(
+                v_airspeed_mat[i, :], angle_of_attack_vec[i], flap_commands[i, :])
             X_aero = np.vstack((X_aero, X_curr))
             aero_features_bar.next()
         aero_features_bar.finish()
         wing_coef_list = ["c_d_wing_xz_offset", "c_d_wing_xz_lin", "c_d_wing_xz_quad", "c_d_wing_xz_stall_min", "c_d_wing_xz_stall_90_deg",
                           "c_l_wing_xz_offset", "c_l_wing_xz_lin", "c_l_wing_xz_stall"]
-        aero_coef_list = wing_coef_list
+        flap_coef_list = ["c_d_ail_lin",
+                          "c_d_ail_quad"]
+        aero_coef_list = wing_coef_list + flap_coef_list
         return X_aero, aero_coef_list
